+++ .github/workflows/register.yaml
@@ -1,4 +1,4 @@
-name: Deploy Flask register App
+name: Deploy Flask Web App
 
 on: workflow_dispatch
 
@@ -29,7 +29,7 @@ jobs:
       id: build-image
       uses: docker/build-push-action@v6
       with:
-        context: ./flask-register-service
+        context: ./flask-webapp-service
         push: true
         tags: |
           ${{ steps.login-ecr.outputs.registry }}/${{ vars.REGISTER_ECR_REPOSITORY }}:${{ github.sha }}
